+++ src/app/http/routes.php
@@ -9,7 +9,8 @@
 Route::controller('login', 'Redooor\Redminportal\App\Http\Controllers\LoginController');
 Route::get('logout', 'Redooor\Redminportal\App\Http\Controllers\LoginController@getLogout');
 
-Route::group(['middleware' => 'Redooor\Redminportal\App\Http\Middleware\Authenticate', 'namespace' => 'Redooor\Redminportal\App\Http\Controllers', 'prefix' => 'admin'], function () {
+Route::group(['middleware' => 'Redooor\Redminportal\App\Http\Middleware\Authenticate',
+              'namespace' => 'Redooor\Redminportal\App\Http\Controllers', 'prefix' => 'admin'], function () {
     Route::get('/', function() {
         return redirect('admin/dashboard');
     });
@@ -30,4 +31,5 @@ Route::group(['middleware' => 'Redooor\Redminportal\App\Http\Middleware\Authenti
     Route::controller('users', 'UserController');
     Route::controller('posts', 'PostController');
     Route::controller('pages', 'PageController');
+    Route::controller('orders', 'OrderController');
 });