+++ src/resources/views/orders/view.blade.php
@@ -1,0 +1,146 @@
+@extends('redminportal::layouts.master')
+
+@section('content')
+    <div class="row">
+        <div class="col-md-12">
+            <ol class="breadcrumb">
+                <li><a href="{{ URL::to('admin') }}">{{ Lang::get('redminportal::menus.home') }}</a></li>
+                <li class="active">{{ Lang::get('redminportal::menus.orders') }}</li>
+            </ol>
+        </div>
+    </div>
+
+    @include('redminportal::partials.errors')
+    
+    <div class="row">
+        <div class="col-md-12">
+            <div class="nav-controls text-right">
+                <div class="btn-group" role="group">
+                @if (count($orders) > 0)
+                <a href="" class="btn btn-default btn-sm disabled btn-text">{{ $orders->firstItem() . ' to ' . $orders->lastItem() . ' of ' . $orders->total() }}</a>
+                @endif
+                <button class="btn btn-default btn-sm" data-toggle="modal" data-target="#export-csv">{{ Lang::get('redminportal::buttons.export_csv') }}</button>
+                {!! HTML::link('admin/orders/create', Lang::get('redminportal::buttons.create_new'), array('class' => 'btn btn-primary btn-sm')) !!}
+            </div>
+            </div>
+        </div>
+    </div>
+
+    @if (count($orders) > 0)
+        <table class="table table-bordered table-striped table-condensed">
+            <thead>
+                <tr>
+                    <th>User</th>
+                    <th>Email</th>
+                    <th>Paid</th>
+                    <th>Payment Status</th>
+                    <th>Transaction ID</th>
+                    <th>Ordered on</th>
+                    <th>Products</th>
+                    <th></th>
+                </tr>
+            </thead>
+            <tbody>
+                @foreach($orders as $order)
+                <tr>
+                    @if ($order->user != null)
+                    <td>{{ $order->user->first_name }} {{ $order->user->last_name }}</td>
+                    <td>{{ $order->user->email }}</td>
+                    @else
+                    <td>User deleted</td>
+                    <td>User deleted</td>
+                    @endif
+                    <td>{{ \Redooor\Redminportal\App\Helpers\RHelper::formatCurrency($order->paid, Lang::get('redminportal::currency.currency')) }}</td>
+                    <td>{{ $order->payment_status }}</td>
+                    <td>{{ $order->transaction_id }}</td>
+                    <td>{{ $order->created_at }}</td>
+                    <td class="table-actions text-center">
+                        <div class="btn-group">
+                            <button type="button" class="btn btn-link dropdown-toggle" data-toggle="dropdown">
+								<span class="glyphicon glyphicon-shopping-cart"></span>
+							</button>
+							<ul class="dropdown-menu pull-right" role="menu">
+                                @foreach ($order->products as $product)
+								<li>
+									<a href="">{{ $product->name }} [{{ $product->sku }}]</a>
+								</li>
+                                @endforeach
+							</ul>
+						</div>
+					</td>
+                    <td class="table-actions text-right">
+                        <div class="btn-group">
+                            <button type="button" class="btn btn-link dropdown-toggle" data-toggle="dropdown">
+								<span class="glyphicon glyphicon-option-horizontal"></span>
+							</button>
+							<ul class="dropdown-menu pull-right" role="menu">
+								<li>
+									<a href="{{ URL::to('admin/orders/delete/' . $order->id) }}" class="btn-confirm">
+										<i class="glyphicon glyphicon-remove"></i>Delete</a>
+								</li>
+							</ul>
+						</div>
+					</td>
+                </tr>
+                @endforeach
+            </tbody>
+        </table>
+        <div class="text-center">
+        {!! $orders->render() !!}
+        </div>
+    @else
+        <div class="alert alert-info">No order found</div>
+    @endif
+    <div id="export-csv" class="modal fade">
+        <div class="modal-dialog">
+            <div class="modal-content">
+                {!! Form::open(array('action' => '\Redooor\Redminportal\App\Http\Controllers\ReportController@postOrders', 'report' => 'form')) !!}
+                <div class="modal-header">
+                    <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span><span class="sr-only">Close</span></button>
+                    <h4 class="modal-title">Export to CSV</h4>
+                </div>
+                <div class="modal-body">
+                    <div class='row'>
+                        <div class="col-md-12">
+                            <div class="form-group">
+                                {!! Form::label('start_date', 'Start Date') !!}
+                                <div class="input-group" id='start-date'>
+                                    {!! Form::input('text', 'start_date', null, array('class' => 'form-control datepicker', 'readonly')) !!}
+                                    <span class="input-group-addon"><span class="glyphicon glyphicon-calendar"></span></span>
+                                </div>
+                            </div>
+                            <div class="form-group">
+                                {!! Form::label('end_date', 'End Date') !!}
+                                <div class="input-group" id='end-date'>
+                                    {!! Form::input('text', 'end_date', null, array('class' => 'form-control datepicker', 'readonly')) !!}
+                                    <span class="input-group-addon"><span class="glyphicon glyphicon-calendar"></span></span>
+                                </div>
+                            </div>
+                            <p class="help-block">Leave the dates empty to download all.</p>
+                        </div>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
+                    {!! Form::submit('Download CSV', array('class' => 'btn btn-primary')) !!}
+                </div>
+                {!! Form::close() !!}
+            </div><!-- /.modal-content -->
+        </div><!-- /.modal-dialog -->
+    </div><!-- /.modal -->
+@stop
+
+@section('footer')
+    <script>
+        !function ($) {
+            $(function(){
+                // For datepicker
+                $( '.datepicker' ).datepicker({ dateFormat: 'dd/mm/yy' });
+                $('.datepicker').css('z-index', '1051'); // make picker on top of modal window
+                $( '.input-group-addon' ).click( function() {
+                    $( this ).parent().find('input').datepicker( "show" );
+                });
+            })
+        }(window.jQuery);
+    </script>
+@stop