+++ tests/controllers/ReportControllerTest.php
@@ -57,4 +57,20 @@ class ReportControllerTest extends RedminTestCase
         $this->assertRedirectedTo('admin/purchases');
         $this->assertSessionHasErrors();
     }
+    
+    /**
+     * Test (Fail): access postOrders with input but no data found
+     */
+    public function testReportOrders()
+    {
+        $input = array(
+            'start_date' => '29/02/2016',
+            'end_date' => '29/02/2016'
+        );
+
+        $this->call('POST', 'admin/reports/orders', $input);
+
+        $this->assertRedirectedTo('admin/orders');
+        $this->assertSessionHasErrors();
+    }
 }