+++ tests/models/OrderModelTest.php
@@ -1,0 +1,30 @@
+<?php namespace Redooor\Redminportal\Test;
+
+use Redooor\Redminportal\App\Models\Order;
+
+class OrderModelTest extends BaseModelTest
+{
+    /**
+     * Contructor
+     */
+    public function __construct()
+    {
+        $model = new Order;
+        $testcase = array(
+            'user_id' => 1,
+            'paid' => 10.99,
+            'transaction_id' => 'UK12345YZ',
+            'payment_status' => 'Completed'
+        );
+        
+        parent::__construct($model, $testcase);
+    }
+    
+    /**
+     * Destructor
+     */
+    public function __destruct()
+    {
+        parent::__destruct();
+    }
+}