+++ CHANGELOG.md
@@ -33,6 +33,8 @@ Looking for RedminPortal for Laravel 4.2? Visit the [v0.1 Branch](https://github
 15. Add Post Model and Controller (issue #73).
 16. Revamp Translation implementation for Category, Media, Module, Portfolio, Product and Promotion (issue #85).
 17. Add Page model and Controller (issue #74).
+18. Save orders for products (issue #51).
+19. Download orders in CSV format.
 
 ### Bug fixes:
 1. Translation to support any language (issue #83).